// src/enums/HttpEnums.ts
export enum HTTP_STATUS_CODE {
    OK = 200,
    OKType = 201,
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    FORBIDDEN = 403,
    NOT_FOUND = 404,
    DATA_CONFLICT = 409, // 数据冲突
    LOG_OUT_OF_DATE = 411,
    USER_EXPERIMENT_DATE = 412,
    INTERNAL_SERVER_ERROR = 500,
    WEAPP_NOT_FOUND_IN_SCHOOL = 420,
    WEAPP_NOT_FOUND_STUDENT_ID = 421,
    WEAPP_CONTACT_ADMIN = 422,
    WEAPP_USER_EXISTENCE = 423,
    WEAPP_ERROR = 455,
    WEAPP_ERROR_STOP = 456,
}

export enum HTTP_STATUS_MESSAGE {
    HTTP_MESSAGE_USER_EXPIRE = '当前用户已过期，请重新登录！',
    HTTP_MESSAGE_TOKEN_ILLEGAL = '你没有权限执行该操作！',
    HTTP_MESSAGE_DATA_NOTFOUND = '无法找到该数据！',
    HTTP_MESSAGE_WRITE_DATA_FAIL = '新增数据失败！',
    HTTP_MESSAGE_USER_NOTFOUND = '账号不存在！',
    HTTP_MESSAGE_USER_ILLEGAL = '账号密码不匹配！',
    HTTP_MESSAGE_OVER_SUBUSER_LIMIT = '无法继续创建账号，当前账号数已达到子帐号上限！',
    HTTP_MESSAGE_USER_DISABLE = '你的账号已被禁用',
    HTTP_MESSAGE_USER_EXISTENCE = '此号码已存在',
    HTTP_MESSAGE_CLASS_CREATED = '此班级已创建',
    HTTP_MESSAGE_CLASS_AND_STUDENT_CREATED = "班级中已存在此学生",
    HTTP_MESSAGE_NO_CLASS = "你不是本班级成员",
    HTTP_MESSAGE_SCHOOL_EXISTENCE = "已经存在该学校,请返回后进行编辑",
    HTTP_MESSAGE_NO_CODE = "验证码错误",
    HTTP_MESSAGE_PRECONDITION_FAILED = "先决条件错误:必填字段为空",
    HTTP_MESSAGE_CONTACT_ADMIN = "该学校没有购买实验，请联系管理员！",
    HTTP_MESSAGE_WEAPP_CONTACT_ADMIN = "该学校未购买实验",
    HTTP_MESSAGE_NOT_FOUND_IN_SCHOOL =  "该学校未找到此学号",
}